[FINCAT-10] Separation of duties, fixed CatSet object persisting issue
